- Fix Reading `EliteArmor` and `VeteranArmor` always return `none` armor - Fix wrong condition when loading `TypeExt` causing fatal error - Add safeguard for `Ext::Canary` so the value wont go beyond `size_t::max` limit - Fix crash cused by `Debug::FreeMouse` when some surface is trying to be blacked without checking if the buffer is actually allocated - FIx various S/L problems that randomly appear if the code recompiled

+++ src/Ext/Convert/Body.h
@@ -12,7 +12,7 @@ public:
 	class ExtData final : public Extension<ConvertClass>
 	{
 	public:
-		static constexpr DWORD Canary = 0xAACAACCC;
+		static constexpr size_t Canary = 0xAACAACCC;
 		using base_type = ConvertClass;
 		static constexpr size_t ExtOffset = 0x178;
